Refactor App move to its component folder && Fix imports for App component
